requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
python-dotenv
pydantic
